app/requirements.txt
Flask==2.0.2
Flask-Bcrypt==0.7.1
marshmallow-sqlalchemy==0.26.1
FLASK-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
marshmallow==3.14.0
Flask-WTF==0.15.1
jwt==1.3.1
SQLAlchemy==1.4.26
requests==2.26.0
mysqlclient==2.0.3
Flask-Migrate==3.1.0
rstr==3.0.0
python-dotenv==0.19.1